+++ Application/Enemy/Enemy.cpp
@@ -259,9 +259,9 @@ void Enemy::TrackPlayerMode()
 	vy = (AllPos.m128_f32[1] - TrackPoint.m128_f32[1]);
 	vz = (AllPos.m128_f32[2] - TrackPoint.m128_f32[2]);
 
-	float v2x = pow(vx, 2);
-	float v2y = pow(vy, 2);
-	float v2z = pow(vz, 2);
+	float v2x = powf(vx, 2.f);
+	float v2y = powf(vy, 2.f);
+	float v2z = powf(vz, 2.f);
 	Length = sqrtf(v2x + v2y + v2z);
 
 	float v3x = (vx / Length) * MoveSpeed;
